requirements.txt
mysql-connector-python==8.0.29
requests==2.27.1
scikit-learn==0.23.2
beautifulsoup4==4.11.1
python-dotenv==0.15.0